requirements.txt
python-telegram-bot[webhooks]==21.0.1
python-dotenv==1.0.0
flask==3.0.0
loguru==0.7.2 